Changes made according to the comments in PR #78
