+++ DashAI/back/metrics/Classification/f1.py
@@ -1,10 +1,13 @@
 import numpy as np
 from sklearn.metrics import f1_score
 
-from DashAI.back.metrics.classification_metric import ClassificationMetric
+from DashAI.back.metrics.classification_metric import (
+    ClassificationMetric,
+    validate_inputs,
+)
 
 
-class F1Score(ClassificationMetric):
+class F1(ClassificationMetric):
     """
     F1 score to classification tasks
     """
@@ -31,18 +34,4 @@ class F1Score(ClassificationMetric):
         if multiclass:
             return f1_score(true_labels, pred_labels, average="micro")
         else:
-            return f1_score(true_labels, pred_labels, average="micro")
-
-
-def validate_inputs(true_labels: list, pred_labels: list):
-    """Validate inputs.
-
-    Parameters
-    ----------
-    true_labels : list
-        True labels
-    pred_labels : list
-        Predict labels by the model
-    """
-    if len(true_labels) != len(pred_labels):
-        raise ValueError("The length of the true and predicted labels must be equal.")
+            return f1_score(true_labels, pred_labels, average="binary")
