+++ DashAI/back/metrics/Classification/precision.py
@@ -1,7 +1,10 @@
 import numpy as np
 from sklearn.metrics import precision_score
 
-from DashAI.back.metrics.classification_metric import ClassificationMetric
+from DashAI.back.metrics.classification_metric import (
+    ClassificationMetric,
+    validate_inputs,
+)
 
 
 class Precision(ClassificationMetric):
@@ -31,18 +34,4 @@ class Precision(ClassificationMetric):
         if multiclass:
             return precision_score(true_labels, pred_labels, average="micro")
         else:
-            return precision_score(true_labels, pred_labels, average="micro")
-
-
-def validate_inputs(true_labels: list, pred_labels: list):
-    """Validate inputs.
-
-    Parameters
-    ----------
-    true_labels : list
-        True labels
-    pred_labels : list
-        Predict labels by the model
-    """
-    if len(true_labels) != len(pred_labels):
-        raise ValueError("The length of the true and predicted labels must be equal.")
+            return precision_score(true_labels, pred_labels, average="binary")
