+++ DashAI/back/metrics/__init__.py
@@ -1,5 +1,5 @@
 # flake8: noqa
 from DashAI.back.metrics.Classification.accuracy import Accuracy
-from DashAI.back.metrics.Classification.f1score import F1Score
+from DashAI.back.metrics.Classification.f1 import F1
 from DashAI.back.metrics.Classification.precision import Precision
 from DashAI.back.metrics.Classification.recall import Recall
