+++ DashAI/back/metrics/classification_metric.py
@@ -7,3 +7,17 @@ class ClassificationMetric(BaseMetric):
     """
 
     _compatible_tasks = ["TabularClassificationTask"]
+
+
+def validate_inputs(true_labels: list, pred_labels: list):
+    """Validate inputs.
+
+    Parameters
+    ----------
+    true_labels : list
+        True labels
+    pred_labels : list
+        Predict labels by the model
+    """
+    if len(true_labels) != len(pred_labels):
+        raise ValueError("The length of the true and predicted labels must be equal.")
